Performance improvements when deserializing tries.
Primarily, this uses the same behaviour as the previous bit
based trie. That is, each trie node is prefixed with its size
(including all of its children). This is read first, and then
if the node is to be excluded, the input stream can skip this
many bytes.

There are also a handful of small changes to try and reduce
string equality checks, reduce the number of upper/lower case
changes, and generally reduce object allocations.