Replace bit-based trie to a String based version.
The bit trie only allowed for up to 32 characters in an
alphabet. It also had some hard assumptions about 26 letters
being present. Finally, it has some hard coded assumptions
about "q" always being followed by a "u".

To properly internationalize this game, all of these assumptions
need to be removed. This is done by removing the smaller, more
efficient bit-based trie in preference to a data structure which
is a tree of Strings.

This will require further performance improvements to be usable.

In the process of removing the bit-based trie, this also tries
to remove anywhere else which depends on bit-shifting. For example
the board was represented by a bitmask, as weas the definition
of whether you can transition from one cell to another. These
have been refactored into more appropriate data structures also.

This adds some test coverage to make sure that everything
still works as expected after this major refactor.

NOTE: This has commented out the KeyboardTracker due to
its dependence on bit shifting. That should be reinstated
in the future.