Adding tests for French dict. Also minor improvements around locale handling.
The `CanTransitionMap` used for testing only worked with the English
alphabet. Now it works with any language (by reading the entire
dictionary and looking at what letters are in it).

I spotted a few places where `toLowerCase()` is called without reference
to a `Locale`. There is still one place this is done in `LexicaView`.