Add Persian dictionary.
Required fixing the trie writer, so that it properly measures the number
of bytes in a unicode string, rather than the number of characters.

Limitation: This excludes any word with a zero width non-joiner (Unicode
character 200C). After a discussion with my Iranian friend, he mentioned
that the simplest solution would be to normalize the dictionary to
include the words with this character, but to exclude the character
itself so that it doesn't end up on the board. This will result in some
words which are broken, but at least they will be able to be shown. In
the best case, we would also unnormalize the words when showing them at
the bottom of the screen and the end of the game. However this will
require a further significant refactoring of the dictionary related code
which I don't have the appetite for right now.