No longer use proguard.
Not only does it make debugging stack traces more difficult, but it also
was removing the dictionary files as they are only reference dynamically
now. Previously they were explicitly referenced so Proguard didn't drop
them. I'm aware we can add rules to prevent this, but the file size
difference will no longer be important in relation to the overall size
of the dictionaries included with the new i18n version.