Presume mandatory letter suffixes (e.g. qu) are saved in the trie.
Before there was some confusion in the code about whether the letter
without the suffix (e.g. "q") is stored in the trie, and we apply the
mandatory suffix when pulling the words back out of the trie, or vice
versa. It is the latter - we strore "qu" in the trie.