Reordered ukranian language so that preference keys line up correctly.
The  and  were not lined up, causing
the Ukranian language to be mistaken for 'English (UK)'.

The correct solution would be to remove the notion of 'UK' and 'US' and replace
them with 'en_GB' and 'en_US' respectively, but that will require much more
testing to make sure that people with old versions don't crash when they open
a new version that doesn't contain 'UK' or 'US' preferences.