Minor fix for Ukranian, missing scrabble point value for з.
While running tests for the Polish language, picked up that the Ukranian
language is missing a letter point value for the letter "з". On closer
inspection, it had accidentally added "3" (the arabic numeral) instead.

This hadn't been picked up before because not every letter in the
alphabet always gets used during the test run, so it is a bit lucky.
I wonder if we can have a test that ensures letter scores exist for all
letters?