Show the score for the past selected word.
Only show the score after the word has been let go, otherwise we give
away to the user whether it is even a word at all or not before they
commit.

Also use a fainter colour for the last selected word if it is not a word
or a previously selected word.