Try to ensure boards with at least 45 words.
Stop after 5 attempts though, because for some languages and board sizes (e.g. large Japanese boards)
it will never get many words, so just cut our losses.

Also don't try forever for performance reasons. This is a naive approach that runs on the UI thread
with no loading feedback to the user, so lets try to keep it lean on slower devices.

In the future, this would be better refactored into a different thread,
with a loading screen shown to the user.

Fixes #124.