Default to Wiktionary when defining words for all languages
Polish doesn't use it, as Polish has a definition service that comes from the wordlist provider.

Preferable to DuckDuckGo because it is more open, and supports multiple languages.
Hopefully there is a nice offshoot that if someone tries to define a word that
doesn't exist in Wiktionary yet (e.g. Ukranian only has on the order of 10,000 words)
then players could add a definition.