Correctly bundle dependencies for language analysis gradle task.
This was broken with an upgrade to the gradle files, whereby the 'compilation' dependencies
were replaced with 'implementation'. There are plenty of workarounds online, but the best
approach seemed to be to use a 3rd party plugin which does this for us. That should future
proof it from futher changes in the future.