Tweak genetic algorithm that produces letter frequencies.
Previously, any letter, no matter how infrequent, would seed with an initial
probability between 1 and 100. That worked fine, because words of low probability
got pushed down quite quickly by the GA. However for some reason the new
German wordlist didn't work like this, and we ended up with many uncommon letters
with quite high probabilities.

This tweaks the initial seeding altorithm so that initial probabilities are chosen
proportional to the frequency of a letter in the wordlist, which can be
toggled with the hard coded PENALISE_INFREQUENT_LETTERS.