Fix issue 126
Diacritics should not be taken into consideration in French. Word games (crossword puzzles, etc) typically ignore diacritics, and this would considerably extend the number of admissible words for any set of letters (in the previous versions, ~100 words are possible on a 5x5 board compared to a few hundred in English, and many letters are unavailable).