Add test to ensure language names used for lookup match the actual language names
This is not an issue presently, but could become one.
Identified when investigating an issue getting the lexicon
on a fresh install, whereby the system locale was used
as the language name instead of the language name.

it was pretty easy to change this to use the language
name instead of system locale, but it also means we
should test to make sure that the language names are
always in line.