Better selecting of default lexicon.
In the past, it assumed that the system locale and language name were the same.
This did not hold true for de_DE and fr_FR.

Instead of using the locale, we now use the language name, which is a stronger
assumption because we are in full control of it.