Remove pt_BR words with k, w, or y in them.
See #243 for discussion on correctness vs fun.

Given they are only used in 295 words out of 628625, it is always
extremely unlikely that you will be able to make a word with any
of these letters if they appear on a board.

Left them in the full, extended dictionary which includes diacritics.