Fix multiplayer games with "Qu"
The previous URL encoding of games fails when a single cell contains
more than one character. Update the encoding to separate the cells with
a comma and base64 encode so that multiple characters work. The minimum
version ('mv') encoded in the URL is used to still accept games from an
old client.

When receiving a URL from an older version with a "Qu" in it, the code
attempts to fix up the encoding by using applyMandatorySuffix() to
discover if there are any additional characters to skip over.

Sadly older versions didn't check the minimum version so won't even
display an error when seeing the new version. Add a check so that at
least in the future if the format changes again a meaningful error can
be given and bump the versionCode.

Fixes #256.