Fix multiplayer games with "Qu"
The previous URL encoding of games fails when a single cell contains
more than one character. Update the encoding to separate the cells with
an underscore so that multiple characters work. The minimum version
('mv') encoded in the URL is used to still accept games from an old
client.

Sadly older versions didn't check the minimum version so won't even
display an error when seeing the new version. Add a check so that at
least in the future if the format changes again a meaningful error can
be given and bump the versionCode.

Fixes #256.