Update ShareQrTest for the new URL scheme
The test parseHumanReadableQR checked that the (hardcoded) URL could be
round-tripped. However the serialize() function will return a new style
URL which doesn't match. Drop this part of the test and rename the test
function to include 'Legacy'.

Also add a new test function for the new style URL which includes a
multicharacter 'Qu' and the round-trip test.

A new 'mock' version of android.util.Base64 is included to ensure that
the tests can pass outside the Android environment.