Dutch letters based on word distributions
The counts are the number of words in the dictionary that contain the letter. The second column is the count for words that contain two of this letter, the third column contains three of the letter, etc.

Closes #307. The previous letter distribution created by the genetic algorithm favoured a subset of letters, resulting in finding the same words in each game. This distribution is more varied and reflects the language better, and results in a more fun game in my opinion.

The fall-off of subsequent columns is quite steep, and this is not optimal for 6x6 boards. Even though words with multiple times the letter w are rare, it wouldn't be strange to have multiple w-s in a 6x6 game. Then you would have multiple options to create words that contain one w. I experimented with increasing subsequent columns with a fraction of the previous column. This gave some improvement (around 10% more words in 6x6 games), but in the end I decided against this because it's more complex, lacks proper motivation, and only gives a small improvement.