Changed Android QR link to lexica://multiplayer
Android QR code is to enable other Android device to scan the code and launch the game without the need to navigate the web page on the Internet.
When there is an Internet connection present, the "Share" option is likely to be used. The QR code is for direct scanning with the app to share the game locally, and it should not require Internet connectivity.