GameDetailsView compatibility with API levels < 21
Fixes #330.

This was introduce in 3356a6a678f555c4d247b5f5c7ee56506a90194d with the
comment that @JvmOverloads ensures no-NewApi compatibility, but that is
not correct: although we correctly expose the 1-args, 2-args and 3-args
constructors for GameDetails view, on API levels < 21, the 4-args
FrameLayout constructor we use does not exist!

According to the Android documentation, the purpose of the 4-args[^1]
and 3-args[^2] constructors is to allow subclasses to use their own base
style when they are inflating, but they are not called by the framework,
which calls the 2-args constructor when inflating from XML[^3].

This patch changes GameDetailsView to use the 2-args constructor (with
@JvmOverloads), following the reasoning outlined
[here](https://proandroiddev.com/misconception-about-kotlin-jvmoverloads-for-android-view-creation-cb88f432e1fe).

[^1]: https://developer.android.com/reference/android/view/View#View(android.content.Context,%20android.util.AttributeSet,%20int,%20int)
[^2]: https://developer.android.com/reference/android/view/View#View(android.content.Context,%20android.util.AttributeSet,%20int)
[^3]: https://developer.android.com/reference/android/view/View#View(android.content.Context,%20android.util.AttributeSet)