Fix multiplayer links to not require internet access
Changes the 'https://lexica.github.io' baseurl for Android link sharing
to be a 'lexica://' link so that there isn't any need to query for a
Digital Asset Links JSON on the web site in order for Android to allow
us to be the handler for such links.

Does rely on the behavior of the QR scanning app to correctly handle a
lexica:// link embedded in the QR code; in my testing the Samsung S22
camera app handles it as expected, but a Pixel 6a didn't and instead
tried to perform a web search for the URI. If need be, we could add a
camera QR scanning feature to Lexica to make it consistent.