Remove FancyButton dependency in favour of MaterialButton.
FancyButton library is no longer supported, but more importantly the
version we were using is not available on MavenCentral. Rather, we
depended on it being in jcenter, which is no more.

Due to this, we are actually unable to build the game. This means not
only can we not release updated translations, etc, but we cannot update
the target SDK as required by Google Play to release new versions there.

Luckily, everything we need is supported via MaterialButton, but the
change is bigger than I would have hoped.