Tweak Espresso tests after refactoring all buttons.
After pulling out the FancyButton dependency, it seemed prudent
to ensure all tests passed. Also, the screenshots haven't been
regenerated for a while, and given that the buttons
resulted in a slight change to the UI, it also seemed like a
good idea to ensure they were all successfully regenerated.

This proved fruitful. No bugs were found, but the tests were
woefully out of date. Also, it did pick up UX issues with
unnecessary word wrapping on some buttons in some locales
which have been addressed in this commit.