Add Finnish language
Add the Finnish dictionary from the governmental linguistic institute of
Finland, available under the CC-BY license. A script for fetching and
generating the dictionary is included.

The letter distribution data was the best in about a dozen runs in terms
of minimum and mean, and second best in terms of score.