Ensure all letters are shuffled correctly (not just the first 15).
In theory, this will have some sort of impact on the number of words
available on larger boards (beyond 4x4), but with a cursory test of
a few board generations in English (UK), it didn't have a large impact.

Even if it does have a large impact, it is still better to do this
right and address consequences as they come, rather than having
a strange corner in the bottom right with less common letters.