Correctly show max score when viewing in-game found words.
There was a performance optimisation to not calculate the word score
for found words twice. It did this by calculating the current score, then
when calculating the max score, only do so for the words which have
not yet been scored.

To make it clearer what is going on, it now just re-counts if neccesary.
It doesn't seem like this performance hit is worth the chance of (well,
the actual) bugs.

Fixes #315.