analyseLanguage gradle tasks write letter probs to disk.
Previously, it could not do this because it ran the GA 5 times,
and it was up to the user to decide which to keep. Thus it wrote
all output to the tmp directory.

Now that we count letter frequencies from dictionaries, it is deterministic
(though the comment about how many letters are on any randomly generated
board differs).